requirements.txt
catt
filelock
Adafruit_SSD1306
Pillow